- remove more mentions of TransformerBase  - move Terrier-specific pipeline rewriting to batchretrieve.py  - refactor to make transformer.py smaller
