+++ pyterrier/debug.py
@@ -1,7 +1,7 @@
-from .transformer import TransformerBase
+from . import Transformer
 from typing import List
 
-def print_columns(by_query : bool = False, message : str = None) -> TransformerBase:
+def print_columns(by_query : bool = False, message : str = None) -> Transformer:
     """
     Returns a transformer that can be inserted into pipelines that can print the column names of the dataframe
     at this stage in the pipeline:
@@ -37,7 +37,7 @@ def print_columns(by_query : bool = False, message : str = None) -> TransformerB
 
 def print_num_rows(
         by_query = True, 
-        msg="num_rows") -> TransformerBase:
+        msg="num_rows") -> Transformer:
     """
     Returns a transformer that can be inserted into pipelines that can print the number of rows names of the dataframe
     at this stage in the pipeline:
@@ -83,7 +83,7 @@ def print_rows(
         jupyter: bool = True, 
         head : int = 2, 
         message : str = None, 
-        columns : List[str] = None) -> TransformerBase:
+        columns : List[str] = None) -> Transformer:
     """
     Returns a transformer that can be inserted into pipelines that can print some of the dataframe
     at this stage in the pipeline:
