+++ pyterrier/ltr.py
@@ -1,6 +1,6 @@
 
-
-from .transformer import TransformerBase, EstimatorBase
+from . import Transformer
+from .transformer import EstimatorBase
 from .apply import doc_score, doc_features
 from .model import add_ranks
 from typing import Sequence, Union
@@ -9,7 +9,7 @@ import numpy as np
 FeatureList = Union[Sequence[int], int]
 
 
-class AblateFeatures(TransformerBase):
+class AblateFeatures(Transformer):
     
     def __init__(self, fids: FeatureList):
         self.fids = fids if isinstance(fids, list) else [fids]
@@ -28,7 +28,7 @@ class AblateFeatures(TransformerBase):
         topics_and_res["features"] = topics_and_res.apply(_reset, axis=1)
         return topics_and_res
 
-class KeepFeatures(TransformerBase):
+class KeepFeatures(Transformer):
     
     def __init__(self, fids : FeatureList):
         self.fids = fids if isinstance(fids, list) else [fids]
@@ -204,7 +204,7 @@ class FastRankEstimator(EstimatorBase):
         test_DF["score"] = scores
         return add_ranks(test_DF)
 
-def ablate_features(fids : FeatureList) -> TransformerBase:
+def ablate_features(fids : FeatureList) -> Transformer:
     """
         Ablates features (sets feature value to 0) from a pipeline. This is useful for 
         performing feature ablation studies, whereby a feature is removed from the pipeline
@@ -215,7 +215,7 @@ def ablate_features(fids : FeatureList) -> TransformerBase:
     """
     return AblateFeatures(fids)
 
-def keep_features(fids : FeatureList) -> TransformerBase:
+def keep_features(fids : FeatureList) -> Transformer:
     """
         Reduces the features in a pipeline to only those mentioned. This is useful for 
         performing feature ablation studies, whereby only some features are kept 
@@ -226,7 +226,7 @@ def keep_features(fids : FeatureList) -> TransformerBase:
     """
     return KeepFeatures(fids)
 
-def feature_to_score(fid : int) -> TransformerBase:
+def feature_to_score(fid : int) -> Transformer:
     """
         Applies a specified feature for ranking. Useful for evaluating which of a number of 
         pre-computed features are useful for ranking. 
@@ -236,7 +236,7 @@ def feature_to_score(fid : int) -> TransformerBase:
     """
     return doc_score(lambda row : row["features"][fid])
 
-def apply_learned_model(learner, form : str = 'regression', **kwargs) -> TransformerBase:
+def apply_learned_model(learner, form : str = 'regression', **kwargs) -> Transformer:
     """
         Results in a transformer that can take in documents that have a "features" column,
         and pass that to the specified learner via its transform() function, to obtain the
@@ -257,7 +257,7 @@ def apply_learned_model(learner, form : str = 'regression', **kwargs) -> Transfo
         return FastRankEstimator(learner, **kwargs)
     return RegressionTransformer(learner, **kwargs)
 
-def score_to_feature() -> TransformerBase:
+def score_to_feature() -> Transformer:
     """
         Takes the document's "score" from the score attribute, and uses it as a single feature. 
         In particular, a feature union operator does not use any score of the documents in the
