+++ pyterrier/text.py
@@ -1,4 +1,4 @@
-from pyterrier.transformer import TransformerBase
+from pyterrier.transformer import Transformer
 from . import Transformer
 from pyterrier.datasets import IRDSDataset
 import more_itertools
@@ -15,7 +15,7 @@ def get_text(
         indexlike, 
         metadata : Union[str,List[str]] = "body", 
         by_query : bool = False,
-        verbose : bool = False) -> TransformerBase:
+        verbose : bool = False) -> Transformer:
     """
     A utility transformer for obtaining the text from the text of documents (or other document metadata) from Terrier's MetaIndex
     or an IRDSDataset docstore.
@@ -124,7 +124,7 @@ def _add_text_irds_docstore(irds_dataset, metadata):
     return add_text_function_docids
 
 
-def scorer(*args, **kwargs) -> TransformerBase:
+def scorer(*args, **kwargs) -> Transformer:
     """
     This allows scoring of the documents with respect to a query, without creating an index first. 
     This is an alias to pt.TextScorer(). Internally, a Terrier memory index is created, before being
@@ -153,7 +153,7 @@ def scorer(*args, **kwargs) -> TransformerBase:
     import pyterrier as pt
     return pt.batchretrieve.TextScorer(*args, **kwargs)
 
-def sliding( text_attr='body', length=150, stride=75, join=' ', prepend_attr='title', **kwargs) -> TransformerBase:
+def sliding( text_attr='body', length=150, stride=75, join=' ', prepend_attr='title', **kwargs) -> Transformer:
     r"""
     A useful transformer for splitting long documents into smaller passages within a pipeline. This applies a *sliding* window over the
     text, where each passage is the give number of tokens long. Passages can overlap, if the stride is set smaller than the length. In
@@ -209,28 +209,28 @@ def sliding( text_attr='body', length=150, stride=75, join=' ', prepend_attr='ti
         **kwargs
     )
 
-def max_passage() -> TransformerBase:
+def max_passage() -> Transformer:
     """
     Scores each document based on the maximum score of any constituent passage. Applied after a sliding window transformation
     has been scored.
     """
     return MaxPassage()
 
-def mean_passage() -> TransformerBase:
+def mean_passage() -> Transformer:
     """
     Scores each document based on the mean score of all constituent passages. Applied after a sliding window transformation
     has been scored.
     """
     return MeanPassage()
 
-def first_passage() -> TransformerBase:
+def first_passage() -> Transformer:
     """
     Scores each document based on score of the first passage of that document. Note that this transformer is rarely used in conjunction with
     the sliding window transformer, as all passages would required to be scored, only for the first one to be used.
     """
     return FirstPassage()
 
-def kmaxavg_passage(k : int) -> TransformerBase:
+def kmaxavg_passage(k : int) -> Transformer:
     """
     Scores each document based on the average score of the top scoring k passages. Generalises combination of mean_passage()
     and max_passage(). Proposed in [Chen2020].
@@ -309,7 +309,7 @@ def snippets(
     return pt.apply.generic(_qbsjoin)   
 
 
-class DePassager(TransformerBase):
+class DePassager(Transformer):
 
     def __init__(self, agg="max", **kwargs):
         super().__init__(**kwargs)
@@ -379,7 +379,7 @@ class MeanPassage(DePassager):
         super().__init__(**kwargs)
 
 
-class SlidingWindowPassager(TransformerBase):
+class SlidingWindowPassager(Transformer):
 
     def __init__(self, text_attr='body', title_attr='title', passage_length=150, passage_stride=75, join=' ', prepend_title=True, **kwargs):
         super().__init__(**kwargs)
